requirements.txt
streamlit==1.32.2
google-generativeai==0.3.2
PyPDF2==3.0.1
python-dotenv==1.0.1
